requirements.txt
pymupdf>=1.24.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
html2text>=2024.2.26
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pyyaml>=6.0.1
click>=8.1.0
python-dotenv>=1.0.1
ftfy>=6.2.0
pillow>=10.0.0
pytesseract>=0.3.10
openai>=1.50.0
pytest>=8.0.0
pytest-cov>=6.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
email-validator>=2.2.0
httpx>=0.27.0
pytest-playwright>=0.5.0
requests-mock>=1.12.1
